Components: Extract a reusable PostPendingStatus component
